+++ .gitignore
@@ -2,3 +2,4 @@
 *~
 venv
 .vscode/launch.json
+__pycache__
