+++ assets/outline.tex
@@ -19,7 +19,7 @@
 \pagestyle{fancy}
 \fancyhf{}
 \lfoot{\textcolor{gray}{Course Outline, Technology \& Innovation}}
-\lhead{{\includegraphics[width=120mm]{scs.png}}}
+\lhead{{\includegraphics[width=120mm]{assets/scs.png}}}
 \rfoot{\textcolor{gray}{!!CODE!! (Section !!SECTION!!),  Page {\thepage}/ \pageref{LastPage}}}
 
 \setstretch{1.3}
