+++ docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3.8'
+services:
+
+    outline:
+        image: outline:1.0.1
+        build:
+          context: .
+          dockerfile: './Dockerfile'
+        ports:
+          - 8000:8000
+        working_dir: /app
+
+        volumes:
+          - .:/app
+
+        command: /bin/bash -c "uvicorn main:app --host  0.0.0.0"
