+++ pdflatexHandle.sh
@@ -1,16 +1,18 @@
 #!/bin/bash
-pdflatex $1
-pdflatex $1
-pdflatex $1
-NAME=`echo "$1" | cut -d'.' -f1`
-EXTENSION=`echo "$1" | cut -d'.' -f2`
-cp $NAME.pdf pdfs/$NAME.pdf
+pdflatex --output-directory=tmp/ $1
+pdflatex --output-directory=tmp/ $1
+pdflatex --output-directory=tmp/ $1
+echo $1
+FILENAME=`echo "$1" | cut -d'/' -f2`
+NAME=`echo "$FILENAME" | cut -d'.' -f1`
+EXTENSION=`echo "$FILENAME" | cut -d'.' -f2`
+cp tmp/$NAME.pdf pdfs/$NAME.pdf
 
-rm $NAME.pdf
-rm $NAME.toc
-rm $NAME.aux
-rm $NAME.log
-rm $NAME.out
-rm $NAME.tex
+rm tmp/$NAME.pdf
+rm tmp/$NAME.toc
+rm tmp/$NAME.aux
+rm tmp/$NAME.log
+rm tmp/$NAME.out
+rm tmp/$NAME.tex
 echo "$NAME done"
