+++ pdfs/.gitignore
@@ -1,1 +1,2 @@
 *.pdf
+*.zip
