+++ requirements.txt
@@ -1,0 +1,39 @@
+anyio==3.6.1
+asgiref==3.5.2
+autopep8==1.7.0
+certifi==2022.6.15
+charset-normalizer==2.1.0
+click==8.1.3
+dnspython==2.2.1
+email-validator==1.2.1
+et-xmlfile==1.1.0
+fastapi==0.79.0
+h11==0.13.0
+httptools==0.4.0
+idna==3.3
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.1
+numpy==1.23.1
+openpyxl==3.0.10
+orjson==3.7.11
+pandas==1.4.3
+pycodestyle==2.9.1
+pydantic==1.9.1
+python-dateutil==2.8.2
+python-dotenv==0.20.0
+python-multipart==0.0.5
+pytz==2022.1
+PyYAML==6.0
+requests==2.28.1
+six==1.16.0
+sniffio==1.2.0
+starlette==0.19.1
+toml==0.10.2
+typing-extensions==4.3.0
+ujson==5.4.0
+urllib3==1.26.11
+uvicorn==0.17.6
+uvloop==0.16.0
+watchgod==0.8.2
+websockets==10.3
